bug fixes and code cleanup

+++ samples/go/noms-perf/ase/iterator.go
@@ -1,11 +1,14 @@
 package ase
 
+// SortedEditItr is a KVPIterator implementation that takes two KVPCollItr and merges them as it iterates
 type SortedEditItr struct {
 	leftItr  *KVPCollItr
 	rightItr *KVPCollItr
 	done     bool
 }
 
+// NewSortedEditItr creates an iterator from two KVPCollection references.  As the iterator iterates it
+// merges the collections and iterates in order
 func NewSortedEditItr(left, right *KVPCollection) *SortedEditItr {
 	leftItr := NewItr(left)
 	rightItr := NewItr(right)
@@ -13,6 +16,7 @@ func NewSortedEditItr(left, right *KVPCollection) *SortedEditItr {
 	return &SortedEditItr{leftItr, rightItr, false}
 }
 
+// Next returns the next KVP
 func (itr *SortedEditItr) Next() *KVP {
 	if itr.done {
 		return nil
@@ -29,6 +33,7 @@ func (itr *SortedEditItr) Next() *KVP {
 	return kvp
 }
 
-func (itr *SortedEditItr) Size() int {
+// Size returns the total number of elements this iterator will iterate over.
+func (itr *SortedEditItr) Size() int64 {
 	return itr.leftItr.coll.totalSize + itr.rightItr.coll.totalSize
 }
